move parts into the src directory
