+++ lib/src/adapter.dart
