+++ lib/src/operations.dart
